week-8/requirements.txt
redshift-connector
pandas
python-dotenv
